requirements.txt
python-telegram-bot[job-queue]
requests==2.32.3
